requirements.txt
absl-py==0.7.1
astor==0.6.2
astroid==1.6.1
attrdict
#azure-mgmt-batchai==2.0.0
#azure==3.0.0
bleach==1.5.0
blindspin
cached-property==1.4.2
csbdeep==0.1.1
daiquiri==1.5.0
decorator==4.2.1
fire
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
imageio==2.3.0
imgaug==0.2.5
isort==4.3.4
#javabridge==1.0.18
Keras==2.2.4
keras-vis==0.4.1
lazy-object-proxy==1.3.1
mahotas==1.4.4
Markdown==2.6.11
marshmallow==3.0.3
matplotlib==2.2.3
mccabe==0.6.1
more-itertools==7.2.0
networkx==2.1
numpy==1.16.0
opencv-python==3.4.0.12
orderedattrdict==1.5.1
pandas==0.24.2
Pillow==5.0.0
protobuf==3.6.1
pyradiomics==2.1.2
python-bioformats==1.4.0
pytest==4.4.0
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.14.2
scikit-learn==0.19.1
scipy==1.0.0
SimpleITK==1.2.0
six==1.11.0
statsmodels==0.9.0
tensorboard==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tifffile==0.15.1
tqdm==4.19.6
webcolors==1.8.1
Werkzeug==0.14.1
wrapt==1.11.2
